fix: use the correct names for pages - req 33
